squash!: Refactor to define admin view permissions in data classes
